Revamp testing framework to support solidity tests.
